Added UI to collect essential info
